+++ packages/sdk-core/src/index.ts
@@ -1,8 +1,3 @@
-export * from './converter';
-export * from './defaultAmountWithDecimals';
-export * from './formatUnitAmount';
-export * from './parseTo18Decimals';
-export * from './reduceBalanceToDenom';
-export * from './units';
+export * from './modules/';
 
 export const sdk = {};
