+++ packages/sdk-core/src/modules/address/index.ts
@@ -1,6 +1,20 @@
-import { hexToU8a, stringToU8a, u8aConcat } from '@polkadot/util';
+import { hexToU8a, isHex, stringToU8a, u8aConcat } from '@polkadot/util';
+import { decodeAddress, encodeAddress, checkAddress } from '@polkadot/util-crypto';
 import { blake2AsU8a } from '@polkadot/util-crypto/blake2';
 import Keyring, { createPair } from '@polkadot/keyring';
+
+export const isValidAddressPolkadotAddress = (address: string, prefix?: number): boolean => {
+  try {
+    if (prefix) {
+      return checkAddress(address, prefix)[0];
+    } else {
+      encodeAddress(isHex(address) ? hexToU8a(address) : decodeAddress(address));
+      return true;
+    }
+  } catch (error) {
+    return false;
+  }
+};
 
 export function evmConverter(evmAddress = ''): string {
   try {
